// models/configuration.ts
export interface Configuration {
    name?: string;
    description?: string;
    appUrl?: string;
    backgroundColor?: string;
    footerColor?: string;
    titleColor?: string;
    footerTextColor?: string;
    footerText?: string;
    favicon: string;
    icon512: string;
    icon192: string;
    icon384: string;
    ogImage: string;
}